Text_to_Handwritten_Text_Converter-master/requirements.txt
Pillow==7.1.2
pywhatkit==4.7